+++ src/main.cli.ts
@@ -1,9 +1,5 @@
 #!/usr/bin/env node
-import { CLIApplication } from './cli/cli-application.js';
-import { HelpCommand } from './cli/commands/help.command.js';
-import { ImportCommand } from './cli/commands/import.command.js';
-import { VersionCommand } from './cli/commands/version.command.js';
-import { GenerateCommand } from './cli/index.js';
+import { CLIApplication, HelpCommand, ImportCommand, VersionCommand, GenerateCommand } from './cli/index.js';
 
 const bootstrap = () => {
   const cliApplication = new CLIApplication();
